# docs/specs/openapi.yaml
openapi: "3.1.0"
info:
  title: Aether Platform API
  description: |
    White-label SaaS коммуникационная платформа.
    Multi-tenant, plugin architecture, AI-powered.

    ## Аутентификация
    - **Bearer JWT**: `Authorization: Bearer <access_token>`
    - **API Key**: `X-API-Key: ak_...`

    ## Tenant isolation
    Все tenant-specific эндпоинты требуют JWT с `organisation_id` claim.
    Admin эндпоинты требуют `is_superadmin = true`.

    ## Rate Limiting
    - Auth: IP-based (5 login/min, 3 signup/hour)
    - Tenant: per-plan (60-6000 RPM)
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    ## Pagination
    Cursor-based: `?cursor=<last_id>&limit=<1-100>`, default 20.
    Response: `{"items": [...], "next_cursor": "uuid"}`.

    ## Errors
    Все ошибки возвращают JSON: `{"error": {"code": "...", "message": "...", "details": {...}}}`
  version: "1.0.0"
  contact:
    name: Aether Platform
    email: dev@aether.cloud
servers:
  - url: https://api.aether.cloud/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
security:
  - bearerAuth: []
  - apiKeyAuth: []

# ============================================================================
# SECURITY SCHEMES
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/refresh
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'API key from /auth/api-keys (prefix: ak_)'

  # ============================================================================
  # COMMON SCHEMAS
  # ============================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "invalid_credentials"
            message:
              type: string
              description: Human-readable error message (localized)
              example: "Неверный email или пароль"
            details:
              type: object
              description: Field-level validation errors
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Ошибка валидации"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # --- Pagination ---
    PaginatedResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {}
        next_cursor:
          type: string
          format: uuid
          nullable: true
          description: Cursor for next page, null if last page
        total:
          type: integer
          description: Total count (when available)
          example: 42

    # --- Auth ---
...
    UserResponse:
      type: object
      required: [id, email, full_name, is_active, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        avatar_url:
          type: string
          format: uri
          nullable: true
        is_active:
          type: boolean
        is_superadmin:
          type: boolean
        mfa_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        locale:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrganisationResponse:
      type: object
      required: [id, name, slug, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logo_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
    MembershipResponse:
      type: object
      required: [id, user_id, organisation_id, role]
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserResponse'
        organisation:
          $ref: '#/components/schemas/OrganisationResponse'
        role:
          type: string
          enum: [owner, admin, member, viewer]
        invited_at:
          type: string
          format: date-time
          nullable: true
        accepted_at:
          type: string
          format: date-time
          nullable: true
    InviteUserRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member, viewer]

    # --- Channels ---
...
  parameters:
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
        format: uuid
      description: Cursor для пагинации (id последнего элемента предыдущей страницы)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Количество элементов на странице

  # ============================================================================
  # COMMON RESPONSES
  # ============================================================================
  responses:
    401Unauthorized:
      description: Не авторизован (JWT истёк или отсутствует)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "unauthorized"
              message: "Требуется аутентификация"
    403Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    404NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    422ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /organisations:
    get:
      operationId: listOrganisations
      summary: Список организаций
      description: Для обычных пользователей — только их организации. Для superadmin — все.
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Список организаций
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrganisationResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
    post:
      operationId: createOrganisation
      summary: Создать организацию (admin)
      tags: [Organisations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 255
                slug:
                  type: string
                  minLength: 3
                  maxLength: 63
                  pattern: '^[a-z0-9-]+$'
      responses:
        '201':
          description: Организация создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /organisations/{org_id}:
    get:
      operationId: getOrganisation
      summary: Детали организации
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали организации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
    patch:
      operationId: updateOrganisation
      summary: Обновить организацию
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 255
                logo_url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Обновлённая организация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
    delete:
      operationId: deleteOrganisation
      summary: Удалить организацию
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Организация удалена
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
  /organisations/{org_id}/members:
    get:
      operationId: listMembers
      summary: Список участников организации
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Список участников
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MembershipResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
  /organisations/{org_id}/invites:
    post:
      operationId: inviteUser
      summary: Пригласить пользователя
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserRequest'
      responses:
        '201':
          description: Приглашение отправлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /organisations/{org_id}/members/{user_id}:
    delete:
      operationId: removeMember
      summary: Удалить участника
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Участник удалён
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
  /organisations/{org_id}/members/{user_id}/role:
    patch:
      operationId: updateMemberRole
      summary: Изменить роль участника
      tags: [Organisations]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member, viewer]
      responses:
        '200':
          description: Роль обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'

  # ==========================================================================
  # API KEYS
  # ==========================================================================
...